+++ kadai1/saza/README.md
@@ -1,13 +1,11 @@
 # 課題 1
 
 ## TODO
-- テストデータの生成を Makefile に追加
-  - 参考: https://github.com/Hiroya-W/gopherdojo-studyroom/tree/kadai1-hiroya-w/kadai1/hiroya-w
-- `make run` コマンドでテスト実行を追加
 - 実装の修正
   - ユーザ定義型を作る
   - main package と分離する
   - 余計なログ出力をやめる
+  - 入力ファイルを出力ファイルで置き換える
 - その他の課題要件
   - Go Modules
   - GoDoc
